Removed leaflet-layervibility file, added it as inline code for the Jenkins build
