+++ src/modules/documentExecution/dashboard/widget/MapWidget/LeafletWrapper.vue
@@ -3,9 +3,8 @@
 </template>
 
 <script lang="ts" setup>
-import { onMounted, onUnmounted, ref, watch } from 'vue'
+import { onMounted, onUnmounted, watch } from 'vue'
 import L from 'leaflet'
-import './leaflet-layervisibility'
 import 'leaflet.markercluster'
 import 'leaflet.markercluster/dist/MarkerCluster.Default.css'
 import 'leaflet/dist/leaflet.css'
@@ -16,6 +15,64 @@ import i18n from '@/App.i18n'
 import { emitter } from '@/modules/documentExecution/dashboard/DashboardHelpers'
 import { clearLayersCache, switchLayerVisibility } from './visibility/MapVisibilityHelper'
 import { IVariable } from '../../Dashboard'
+
+//#region inlined leaflet-layervisibility
+function validateFilter(filterFunc) {
+    if (typeof filterFunc !== 'undefined' && typeof filterFunc === 'function') {
+        return filterFunc
+    }
+    return function () {
+        return true
+    }
+}
+
+function setLayerDisplayStyle(value, context) {
+    if (context._el) context._el.style.display = value
+    else if (context.getElement()) context.getElement().style.display = value
+    return context
+}
+
+function setLayerGroupVisibility(mode, filter, context) {
+    const filterFunc = validateFilter(filter)
+    context.eachLayer((layer) => (filterFunc(layer) ? layer[mode]() : null))
+}
+
+L.Layer.include({
+    hide() {
+        return setLayerDisplayStyle('none', this)
+    },
+    show() {
+        return setLayerDisplayStyle('', this)
+    },
+    isHidden() {
+        return this.getElement().style.display === 'none'
+    },
+    toggleVisibility() {
+        return this.isHidden() ? this.show() : this.hide()
+    }
+})
+
+L.LayerGroup.include({
+    hide(filter) {
+        setLayerGroupVisibility('hide', filter, this)
+    },
+    show(filter) {
+        setLayerGroupVisibility('show', filter, this)
+    },
+    isHidden() {
+        return this.getLayers().every((layer) => layer.isHidden())
+    }
+})
+
+L.Marker.include({
+    hide() {
+        return L.Layer.prototype.hide.call(this, null)
+    },
+    show() {
+        return L.Layer.prototype.show.call(this, null)
+    }
+})
+//#endregion
 
 const appStore = useAppStore()
 const { t } = i18n.global
