--- src/modules/documentExecution/dashboard/widget/MapWidget/leaflet-layervisibility.d.ts
@@ -1,1 +1,0 @@
-declare module './leaflet-layervisibility.js'
