+++ tsconfig.json
@@ -1,6 +1,5 @@
 {
     "compilerOptions": {
-        "allowJs": true,
         "target": "ESNext",
         "module": "esnext",
         "strict": true,
